Test out a footer nav that doesn’t hide itself
In further browser testing, it looks like a lot of the self-hiding nav behavior I’ve seen is actually the browser hiding a site’s bottom nav by increasing the height of the viewport when the user scrolls up. When the browser does this, it automatically frees up screen real estate for the browser’s floating action buttons.